+++ navigation/AppNavigator.tsx
@@ -12,6 +12,9 @@ import BookingScreen from '../screens/BookingScreen';
 import LiveTrackScreen from '../screens/LiveTrackScreen';
 import Bsse7aScreen from '../screens/Bsse7aScreen';
 import LogoShowcaseScreen from '../screens/LogoShowcaseScreen';
+import AvatarShowcaseScreen from '../screens/AvatarShowcaseScreen';
+import QRScannerShowcaseScreen from '../screens/QRScannerShowcaseScreen';
+import TypographyShowcaseScreen from '../screens/TypographyShowcaseScreen';
 
 
 // Create the stack navigator
@@ -69,6 +72,18 @@ const AppNavigator = () => {
         <Stack.Screen
           name="LogoShowcaseScreen"
           component={LogoShowcaseScreen}
+        />
+        <Stack.Screen
+          name="AvatarShowcaseScreen"
+          component={AvatarShowcaseScreen}
+        />
+        <Stack.Screen
+          name="QRScannerShowcaseScreen"
+          component={QRScannerShowcaseScreen}
+        />
+        <Stack.Screen
+          name="TypographyShowcaseScreen"
+          component={TypographyShowcaseScreen}
         />
       </Stack.Navigator>
     </NavigationContainer>